third_party/blink/web_tests/external/wpt/service-workers/service-worker/interface-requirements-sw.https-expected.txt

This is a testharness.js-based test.
[FAIL] Event constructors
  assert_equals: FetchEvent.isReload should not exist expected (undefined) undefined but got (boolean) false
Harness: the test ran to completion.
